Add Bash support to the dotfiles
I tend to not use Bash that often, it's far from my favorite shell,
but bash is not available everywhere, and a lot of systems default
to bash. This will allow me to _at least_ be able to use the system
until I get ZSH installed and my shell changed over. A bonus is that I
don't _need_ to change my shell if I can't or don't want to. The bashrc
will automatically swap shells to zsh if it's available and we're in a
real login shell.