Add Tmux configuration
TMUX is a very new tool in my workflow. It's a terminal multiplexer that
allows me to setup elaborate shell splits (with tabs even!) and then
detach from them and reconnect from a seperate machine. This means I can
work on a machine remotely and pick up where I left on when I get back
to work. It also will allow me to reattach if I have bad connection.