Normalize vim controls and set background
By default, they use jkl; to navigate the windows open on the system,
however I'm wired for the default hjkl. I moved things around to work
better with this layout and to go ahead and disable stacking and use the
same splitting as vim as well.

I also made some changes related to rofi and background settings to make
things a little more usable.