Remove installing nvim from the dotfiles
I've thought about this for a while. The thing is- each system manages
it's packages separately, and ultimately packages should be installed
using those package managers. Importing appimage should be a last
resort, not a first resort. Likely, my packages should be in
documentation, or through a stateful system like puppet.