Add rules for Java and XML development
These are the not the standard rules that I would use when writting
code, however these seem to be Chris's defaults for his Java and I'm
attempting to not rock the boat with the in-house java Development.

Ideally, the long-term solution will be to find a plugin for Neovim
that will auto-adjust the tab/space settings to match the files existing
contents.