Add wezterm configuration
I've started experimenting with wezterm. For the most part I like it,
and I'll likely stick with it for the time being. My ultimate goal is to
have wezterm replace my terminal everywhere- though it'll take a while
before I can have the same level of Guake I've come to expect from
iTerm2 on macOS