Replace OhMyZsh! with Antidote
There are a lot of really nice features that come out of the box with
OhMyZsh, but I used effectively none of them. Switching to Antidote
means that I can have the same features, maybe even more, without the
weight of OhMyZSH. This also means that I do have to do some of the
lifting that OhMyZSH was doing for me, but this is only a few more lines
in my ZSHRC and ultimately results in a faster more explicit experience.

Now if only I could fix the highlight issue with WezTerm...