Massive upgrade for neovim and vim
This updates neovim to leverage the latest LSP and telescope.

I also added a proper vimrc for when I have to fall back to vim, and
I added a better integration for intellij vim. Includes some keybindings
to solve some long-standing issues I've had with copy, cut, and paste
registers.

There's more as well... but you can look through the changes to see what
it is.