Linting + Conditionally configure available software
I walked through each of the roles and added a "Check" condition to
determine if the software is currently installed, and if it isn't,
install it if install_software uses it.

Updated change_when to actually detect when a configuration /
application changed instead of just ignoring it. So changed is truely
changed now.

I also fixed a mountain of linting errors.