shell: Fix XDG_CACHE_HOME issue with Antidote
I made the assumption that the XDG_CACHE_HOME would always be in the
same place and have the same values for shared operating systems- but
this isn't actually a safe assumption to make. Now I generate a plugins
file per-host, which should be a healthy way to do this... assuming we
don't run into additional issues with the files getting cleaned up.