neovim: Bump lsp-zero to the new branch, update Copilot, spelling, add UndoTree
Three major changes have happened here from the developer experience
side of things. Copilot's suggestions are now tucked into the completion
menu, words that Neovim (heh, it thinks Neovim is misspelled) thinks are
misspelled are now underlined, and finally the inclusion of the UndoTree
which allows me to navigate changes in a file through a tree of saves.